Robust depth image sampling
-Make image sampling robust, by taking n number of samples and selecting the ones that are most similar and dropping others.

Cropped function now resamples the  generated points to make a single profile by
setting the axis coordinate value that has lowest deviation to the mean of all points in that axis. This will make deviation to 0 for that particular axis. Whenever robot is moving, it always moves in 2 axes, if it moves in x and z axes that means there should be ideally 0 deviation in y axis. so, setting all values of y to average y value will give us 0 deviation in y axis and robot won't shake.

-set visualization views