Modded Simulation Camera script
*Trim edges in cropped_PC function, by adjusting the bbox bounds.

-overhauled generate_coordinates function: This function now tries to deal with coordinate anomalies which used to generate ghost targets.

It looks at x,y,z coordinates and observes if series is increasing or decreasing plus if the top coordinate is bigger than last one, based on that, it decides whether to delete it or not. Also removes duplicate coordinates from it.

*Added function filter_generate_coordinates:
This function will sample generated coordinates and compare to find the ones with anomalies and discard them. Can choose number of samples to take.