Added a manual offset variable
Added a manual offset variable  to adjust position of the object with respect to the robot's base.
If the object is at same level as the robot's base, then set to 0 else, set to how far it is from the robot's base. Useful in case of vertical scanning or if the robot's base is at higher position than the object.