Ported GUI code to ROS Simulation Camera code
GUI code added to the Sim code.
GUI is fully functional and includes three modes

*Replay mode:
Used to re run the saved coordinates. User can add new offsets to the existing targets and save again.
Can be useful in generating target files where object is symmetric and doesn't require repeated scanning.

*Manual mode:
User had full control of entire process. User is presented with the option to set the initial pose of the robot with a set of predefined poses. User can choose to add offsets between the object and camera if required.

User then needs to select the correct cluster to generate the correct profile and the robot moves to the generated targets in order. Finally, user is presented with various save options where it can either create new file, overwrite or append targets or an existing file.

*Auto mode:
This mode does everything that manual mode do, plus it has the option to pre-select various parameters that user earlier set manually such as selecting cluster, default profile path, number of iterations, save settings, Hide previews etc...
This mode can be useful in most cases where the object is larger and can't fit in the field of view of the camera in one go. User needs to estimate the number of iterations required and the robot will scan the object initially and after completing first motion, it will start with the second iteration from that point onwards.

*Settings:
All the default parameters are configurable and are stored in a separate file. Parameters can be set and tuned from settings according to the use case.